+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Fix Invoke-Pester -OutputXml usage of relative paths 
   - Remove Validate-Xml function
   - Remove legacy object adaptations support
+  - Remove tests testing usage of the global scope
 
 ## 3.0.0-beta2 (July 4, 2014)
   - Add code coverage [GH-148]